support CachePolicy resources (#4)
Adds support for CachePolicy resources. In generating and writing test cases for the CloudFront API, I learned a number of things, including that all CloudFront UPDATE and DELETE APIs require the caller to pass in an `IfMatch` field value equal to the resource's `ETag` field value from a previous GET. If you don't do this, you will get an `InvalidIfMatchVersion` error returned from CloudFront.

Therefore, all of the custom hook points in the generator are for dealing with this issue.

Issue aws-controllers-k8s/community#249

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.