Apply ETag and Quantity inputs
Added hooks for setting the ETag and Quantity fields in the Cloudfront API. This is to prevent the ETag from getting out of sync after a creation or modification of the AWS entity. Additionally, a Quantity field is used for OriginSSLProtocols which can be automatically set by counting the number of items.